extract validation class
