+++ lib/primer/classify.rb
@@ -7,14 +7,13 @@ require_relative "classify/functional_border_colors"
 require_relative "classify/functional_text_colors"
 require_relative "classify/grid"
 require_relative "classify/utilities"
+require_relative "classify/validation"
 
 module Primer
   # :nodoc:
   class Classify
     # Keys where we can simply translate { key: value } into ".key-value"
     CONCAT_KEYS = %i[text box_shadow].freeze
-
-    INVALID_CLASS_NAME_PREFIXES = Regexp.union(["bg-", "color-", "text-", "box-shadow-"] + CONCAT_KEYS.map { |k| "#{k}-" }).freeze
 
     COLOR_KEY = :color
     BG_KEY = :bg
@@ -113,10 +112,6 @@ module Primer
         extracted_results
       end
 
-      def invalid?(class_name)
-        class_name.start_with?(INVALID_CLASS_NAME_PREFIXES) || Primer::Classify::Utilities.supported_selector?(class_name)
-      end
-
       private
 
       def validated_class_names(classes)
@@ -125,7 +120,7 @@ module Primer
         if force_system_arguments? && !ENV["PRIMER_WARNINGS_DISABLED"]
           invalid_class_names =
             classes.split(" ").each_with_object([]) do |class_name, memo|
-              memo << class_name if invalid?(class_name)
+              memo << class_name if Primer::Classify::Validation.invalid?(class_name)
             end
 
           raise ArgumentError, "Use System Arguments (https://primer.style/view-components/system-arguments) instead of Primer CSS class #{'name'.pluralize(invalid_class_names.length)} #{invalid_class_names.to_sentence}. This warning will not be raised in production. Set PRIMER_WARNINGS_DISABLED=1 to disable this warning." if invalid_class_names.any?
