+++ lib/primer/classify/validation.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+module Primer
+  class Classify
+    # :nodoc:
+    class Validation
+      INVALID_CLASS_NAME_PREFIXES = /bg\-|color\-|text\-|box\-shadow\-|text\-|box_shadow\-/.freeze
+
+      class << self
+        def invalid?(class_name)
+          class_name.start_with?(INVALID_CLASS_NAME_PREFIXES) || Primer::Classify::Utilities.supported_selector?(class_name)
+        end
+      end
+    end
+  end
+end
