+++ lib/rubocop/cop/primer/primer_octicon.rb
@@ -2,6 +2,7 @@
 
 require "rubocop"
 require "primer/classify/utilities"
+require "primer/classify/validation"
 
 # :nocov:
 module RuboCop
@@ -53,7 +54,7 @@ module RuboCop
           # check if classes are convertible
           if classes.present?
             system_arguments = ::Primer::Classify::Utilities.classes_to_hash(classes)
-            invalid_classes = (system_arguments[:classes]&.split(" ") || []).select { |class_name| ::Primer::Classify.invalid?(class_name) }
+            invalid_classes = (system_arguments[:classes]&.split(" ") || []).select { |class_name| ::Primer::Classify::Validation.invalid?(class_name) }
 
             # Uses system argument that can't be converted
             return if invalid_classes.present?
