+++ test/lib/classify/validation_test.rb
@@ -1,0 +1,21 @@
+# frozen_string_literal: true
+
+require "test_helper"
+
+class PrimerClassifyValidationTest < Minitest::Test
+  def test_invalid?
+    assert_invalid("d-block")
+    assert_invalid("color-text-primary")
+    assert_valid("custom-class")
+  end
+
+  private
+
+  def assert_invalid(class_name)
+    assert(Primer::Classify::Validation.invalid?(class_name))
+  end
+
+  def assert_valid(class_name)
+    refute(Primer::Classify::Validation.invalid?(class_name))
+  end
+end
